LEC-112 update: Fixed UI on Profile page where it was rendering twice
